requirements.txt
# Telegram
python-telegram-bot==21.6

# HTTP client (async)
httpx==0.27.2

# Scheduler
APScheduler==3.10.4

# Database
SQLAlchemy==2.0.35
aiosqlite==0.20.0

# Charts
matplotlib==3.9.2
plotly==5.24.1
kaleido==0.2.1        # plotly PNG export

# Environment
python-dotenv==1.0.1

# Utilities
pytz==2024.2